Use an interning table to compare and hash FileBasedSource objects.
Speeds up analysis by about 8%.

Note: the interning table uses both the file path and the URI to
distinguish FileBasedSource objects.  This should eliminate the
strange behaviors we would see in the rare case where a single URI
would get associated with multiple file paths (see dartbug.com/22555).

The interning table is never flushed over the lifetime of the
application.  In practice this should not be a problem, because it
only contains one entry per filename/URI, and the total number of
filenames/URIs analyzed during the course of execution should always
be small.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//954063002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44026 260f80e4-7a28-3924-810f-c04153c831b5