Move the dart_precompiled function inside scoping function.
Also make it a property of init.

This is also done for deferred hunks, allowing deferred loading to work
in CSP mode.

Long term I would like the precompiled function to go away, and move
the generated functions into the library-initializer when they are
 needed.

BUG=16898
R=floitsch@google.com

Review URL: https://codereview.chromium.org//968163002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44162 260f80e4-7a28-3924-810f-c04153c831b5