dart2js: Allow to encode side-effects in the spec-string of JS and similar built-ins.
This patch prepares a shift towards two builtins. We currently have ~20 foreign functions that must be caught by the resolver, type-inferrer, builder, ...
In the future there will be only two:
JS(...), and JS_BUILTIN(...)

The JS_BUILTIN function would take a string or enum as second argument (after the spec-string) which would be shared between the runtime and the compiler.
This makes it easier to ensure that no builtin is forgotten in the places where they need to be handled.

This patch also gives more flexibility to the implementor.
For example `JS('bool', 'typeof self["foo"] != "undefined"')` would currently trigger a "depends on index store". With this patch the developer could override the automatic side-effect analysis and force the dependency to be empty.

Note: in theory we could make "JS" a BUILTIN itself, but, given its importance, I prefer to keep it separate.

Note2: with the "ForceInline" CL [0] we can keep the old functions (like "JS_GET_FLAG") and just forward to JS_BUILTIN:

@ForceInline()
bool JS_GET_FLAG(String name) {
  // Given: a shared enum `Builtin`.
  return JS_BUILTIN('bool', Builtin.getFlag, name);
}

[0] https://chromiumcodereview.appspot.com/962703004/

R=herhut@google.com, sra@google.com

Review URL: https://codereview.chromium.org//969093002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44229 260f80e4-7a28-3924-810f-c04153c831b5