Streamline the CPS IR Visitor interface.
Previously, the visitor interface combined two things: (1) the
interface for visiting concrete IR instruction classes and (2) a
default implementation that called the visit function for the
superclass explicitly (but not through super).  For most classes that
extended Visitor this inherited implementation was used to throw an
exception if a visit method for a new class was forgotten.

Instead, it's better to make this class purely an interface which will
give a static warning when a visit method is forgotten.  This change
forces us to implement methods that previously had an inherited (not
necessary correct) implementation.

BUG=
R=asgerf@google.com, karlklose@google.com

Review URL: https://codereview.chromium.org//979693003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44235 260f80e4-7a28-3924-810f-c04153c831b5