JSArray has its own iterator
This replaces many uses of ListIterator with ArrayIterator.
ListIterator tends to become very generic in real programs
(see http://dartbug.com/22671)

There is a tradeoff here between increasing the polymorphism at the moveNext() / get current site and reducing it at the internal sites in ListIterator.
Experiments show that true polymorphic for-in loops, which are quite slow, are a few percent faster, but fake polymorphic loops (i.e. always a JSArray at runtime) and lightly polymorphic loops (usually JSArray) are about twice as fast.

Even if we implemented full optimization of loops (see http://dartbug.com/14952) this change is worthwhile for the compile-time polymorphic cases.

This seems to hold for both d8 and jsshell.

R=lrn@google.com

Review URL: https://codereview.chromium.org//976203002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44263 260f80e4-7a28-3924-810f-c04153c831b5