Add an analyzer option to properly check ".call" methods.
Enabling this option fixes issue 21938.  For now the option is
disabled by default, since there is a lot of code in the dart SDK (and
in third_party) that isn't warning clean with this option enabled.
Once the SDK and third_party code have been cleaned up, we should be
able to turn this option on permanently.

BUG=dartbug.com/21938
R=scheglov@google.com

Review URL: https://codereview.chromium.org//994253002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44390 260f80e4-7a28-3924-810f-c04153c831b5