Don't run dart_style/test/formatter_test on slow platforms
Ivan/Bob: Do we want to simply run these exclusively on the pub/pkg bots? This will simplify rolling deps for this and be similar to what we used to do for pkg/ in the repo.
If so, do we want to remove it from the default selector as well? (i.e., people will not run third_party/pkg_tested unless explictly requesting to, the pub/pkg bots would do that)

R=iposva@google.com, rnystrom@google.com
BUG=

Review URL: https://codereview.chromium.org//1002663002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44423 260f80e4-7a28-3924-810f-c04153c831b5