Make Completer.complete handle a Future with a misbehaving "then" call.
Instead of throwing and getting into an invalid state (isChained but won't ever
complete), the synchronous error from trying to call "then" is used to complete
the future.

This should never happen in practice - it requires a "then" implementation that
is missing or throwing even when given valid arguments.

R=sgjesse@google.com

Review URL: https://codereview.chromium.org//1025133004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44629 260f80e4-7a28-3924-810f-c04153c831b5