Rename MethodInvocation.period to MethodInvocation.operator.
The name is inaccurate--if the method invocation is part of a cascade,
it is the ".." operator.  It's going to get even more inaccurate when
null-aware operators are added, since it will also be used to
represent "?.".

Deprecated getters and setters named "period" are added to avoid
breaking the analyzer API.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//1039713002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44744 260f80e4-7a28-3924-810f-c04153c831b5