Add a core SDK changelog.
This is based on the archive of release notes. Its primary goal is to
provide a canonical source for Dart package developers to refer to
when determining what SDK constraints to use for their packages. It
only goes back through 1.7.0 because older SDKs aren't in wide use and
it's unlikely that package authors will need to know what features
changed back then.

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1064493002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44899 260f80e4-7a28-3924-810f-c04153c831b5