Mostly fix signal-forwarding for "pub run".
When the user presses Control-C on the terminal, it sends SIGINT to
all processes in the process tree. However, this is *not* the case for
other signals or signals sent via other means. As a compromise, "pub
run" now manually forwards all signals other than SIGINT.

R=rnystrom@google.com

Review URL: https://codereview.chromium.org//1060783002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@44902 260f80e4-7a28-3924-810f-c04153c831b5