tree-ir bugfix: Count unseen uses instead of seen uses.
Keeping track of the number of "seen uses" of a variable is too fragile
since it needs bookkeeping whenever we merge two variable uses.

I've changed it to track the number of unseen uses instead. The
bookkeeping is simpler because we aren't going to transform an unseen
use.

BUG=
R=kmillikin@google.com

Review URL: https://codereview.chromium.org//1080343003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@45207 260f80e4-7a28-3924-810f-c04153c831b5